requirements.txt
#
# Requirements for Owlmind

discord==2.3.2
python-dotenv==1.0.1
requests>=2.28.0
audioop-lts; python_version>='3.13'
pymongo==4.6.1
motor==3.3.2
pdfminer.six==20231228
matplotlib==3.8.2
numpy==1.26.3
httpx==0.25.2
python-dateutil==2.8.2
beautifulsoup4==4.12.2
aiofiles==23.2.1
pillow==10.2.0

